rubah file layout untuk perbaikan page
