+++ src/app/layout__.js
@@ -1,17 +1,12 @@
-import './globals.css'
-import { Inter } from 'next/font/google'
-
-const inter = Inter({ subsets: ['latin'] })
-
 export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'Next.js',
+  description: 'Generated by Next.js',
 }
-
+ 
 export default function RootLayout({ children }) {
-  return (
+ return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body>{children}</body>
     </html>
   )
 }
